chore: 카드 설정 페이지 로직 변경
