+++ src/screens/SettingScreen/CardListScreen.tsx
@@ -1,82 +1,19 @@
-import {useState} from 'react';
-import {StyleSheet} from 'react-native';
-import {cardUsagePeriodOptions} from 'constants/screens/SettingScreen';
+import {View} from 'react-native';
 import ScreenContainer from 'components/@common/ScreenContainer';
-import CommonSelectItem from 'components/@common/CommonSelectItem';
-import InputBottomSheet from 'components/@common/InputBottomSheet';
-import CardSettingFixedButton from 'components/screens/SettingScreen/CardSettingFixedButton';
-import SelectForm from 'components/@common/SelectForm';
+import Button from 'components/@common/Buttons/Button';
+import {useNavigation} from '@react-navigation/native';
 
 export default function CardListScreen() {
-  const [cardName, setCardName] = useState('');
-  const [cardInputText, setCardInputText] = useState('');
-  const [isCardNameBottomSheetOpen, setIsCardNameBottomSheetOpen] =
-    useState(false);
-  const [cardUsagePeriodList, setCardUsagePeriodList] = useState(
-    cardUsagePeriodOptions ?? [],
-  );
-  const [keyNoteText, setKeyNoteText] = useState('');
-  const [keyNoteInputText, setKeyNoteInputText] = useState('');
-  const [isKeyNoteBottomSheetOpen, setIsKeyNoteBottomSheetOpen] =
-    useState(false);
+  const navigation = useNavigation();
 
-  const handlePressCardNameSelectItem = () =>
-    setIsCardNameBottomSheetOpen(true);
-
-  const handlePressKeyNoteSelectItem = () => {
-    setIsKeyNoteBottomSheetOpen(true);
-  };
+  const handlePressAddCardButton = () => navigation.navigate('AddCard');
 
   return (
     <ScreenContainer
-      contentContainerStyle={cardSettingScreenStyles.contentContainer}
       fixedBottomComponent={
-        <CardSettingFixedButton cardName={cardName} keyNoteText={keyNoteText} />
+        <Button onPress={handlePressAddCardButton}>카드 추가</Button>
       }>
-      <CommonSelectItem
-        label={'카드명'}
-        handlePressSelectItem={handlePressCardNameSelectItem}
-        value={cardName}
-        placeholder="카드명을 입력해주세요"
-        bottomSheet={
-          <InputBottomSheet
-            isBottomSheetOpen={isCardNameBottomSheetOpen}
-            setIsBottomSheetOpen={setIsCardNameBottomSheetOpen}
-            placeholder="카드명을 입력해주세요"
-            inputText={cardInputText}
-            setText={setCardName}
-            setInputText={setCardInputText}
-          />
-        }
-      />
-      <SelectForm
-        label={'사용 기간'}
-        placeholder="사용 기간을 선택해주세요"
-        categoryList={cardUsagePeriodList}
-        setCategoryList={setCardUsagePeriodList}
-      />
-      <CommonSelectItem
-        label={'적요'}
-        handlePressSelectItem={handlePressKeyNoteSelectItem}
-        value={keyNoteText}
-        placeholder={'적요를 입력해주세요'}
-        bottomSheet={
-          <InputBottomSheet
-            isBottomSheetOpen={isKeyNoteBottomSheetOpen}
-            setIsBottomSheetOpen={setIsKeyNoteBottomSheetOpen}
-            placeholder="적요를 입력해주세요"
-            inputText={keyNoteInputText}
-            setText={setKeyNoteText}
-            setInputText={setKeyNoteInputText}
-          />
-        }
-      />
+      <View />
     </ScreenContainer>
   );
 }
-
-const cardSettingScreenStyles = StyleSheet.create({
-  contentContainer: {
-    gap: 10,
-  },
-});
